requirements.txt
nb_utils
loguru

numpy
opencv-python
pillow
scikit-learn

pandas
numba